task-01 ready

+++ shell/shell-2/task-01.sh
@@ -4,9 +4,9 @@ echo Give the name of the file you want to delete:
 
 read FILENAME
 
-if [[ -f $FILENAME ]]
+if [[ -f $FILENAME && ! -s $FILENAME ]]
 then
-    echo "$FILENAME is a regular file"
+    echo "$FILENAME is a regular file and empty"
     rm $FILENAME
     echo "File $FILENAME deleted"
 else
